Move the code into libraries for reuse
